requirements.txt
opencv-python==4.11.0
Pillow==10.4.0
imagehash==4.3.1
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3